\name{staRt-package}
\alias{staRt-package}
\docType{package}
\title{
Inferenza classica con TI-83 Plus
}
\description{
Una libreria per utilizzare con semplicità le tecniche di
statistica inferenziale presenti sulla calcolatrice scientifica
grafica TI-83 Plus
}
\details{
\tabular{ll}{
Package: \tab staRt\cr
Type: \tab Package\cr
Encoding: \tab latin1\cr
Depends: \tab R, stats, utils, tools\cr
Version: \tab 1.1.12\cr
Date: \tab 2007-10-12\cr
License: \tab GPL-2\cr
}

Index:
\preformatted{
ANOVA                   Analisi della varianza ad una variabile
Chi2.Test               Verifica chi quadrato per tabelle a 2 variabili
LinRegTTest             Verifica t della pendenza della regressione
One.PropZInt            Intervallo di confidenza di 1 proporzione
One.PropZTest           Verifica di una proporzione
One.Var.Stats           Statistiche ad 1 variabile
T.Test                  Verifica di un singolo mu, sigma non nota
TInterval               Intervallo di confidenza di 1 mu, sigma non
                        nota
Two.PropZInt            Intervallo di confidenza per la differenza tra
                        2 proporzioni
Two.PropZTest           Verifica di confronto di 2 proporzioni
Two.SampFTest           Verifica di confronto di 2 sigma
Two.SampTInt            Intervallo di confidenza per la differenza di 2
                        mu, sigma non note
Two.SampTTest           Verifica di confronto di 2 mu, sigma non note
Two.SampZInt            Intervallo di confidenza per la differenza di 2
                        mu, sigma note
Two.SampZTest           Verifica di confronto di 2 mu, sigma note
Two.Var.Stats           Statistiche a 2 variabili
Z.Test                  Verifica di un singolo mu, sigma nota
ZInterval               Intervallo di confidenza di 1 mu, sigma nota
}
}
\author{
Fabio Frascati

Maintainer: Fabio Frascati <fabiofrascati@yahoo.it>
}
\keyword{ package }
